fix: AtomicInteger -> AtomicLong

+++ src/main/java/com/econovation/hellstudy/common/helper/AtomicUtils.java
@@ -1,13 +1,14 @@
 package com.econovation.hellstudy.common.helper;
 
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class AtomicUtils {
 
     // dataBase의 Auto-Increment
-    public static int autoIncrement(AtomicInteger counter) {
-        int value = counter.get();
-        int newValue = value + 1;
+    public static long autoIncrement(AtomicLong counter) {
+        long value = counter.get();
+        long newValue = value + 1;
         counter.set(newValue);
         return newValue;
     }
